refactor: clean the main.py and add some functions to utils
I put the atena_upload and atena_connect into the utils, but in the future this should be in a class for the connections
